Initial commit of Distro example
This commit is based on the (obsolete) hello_android example (now
hello_services). It uses V23 for RPC and mount table globbing for
discovery (since global discovery is not exposed in Java and local
discovery works only through mDNS right now). It does not actually cast
components yet, but presents a list of (globally) available cast
targets.

This commit is subject to v.io/i/1356, and the VDL Java generation is
commented out.

Change-Id: Iba5116324109c48d7ab009d3d3759f6bc469d342