feat(mdtest run/auto): unlock devices if necessary and uninstall testing apps after test execution.  support 'screen-size' property in test spec.
Wake up and unlock the device if it is locked.  This is good for
integration testing, where the devices are locked after not being
used for a long time.  Uninstall tested apps after test execution.
This feature is supported to avoid too many apps running in
background if a device is used to test multiple different apps.

Also add screen-size support in test spec.

Change-Id: Id0f9a209b2da15d944fc383442fc804789974118