feat(mdtest run/auto): add test spec validation checker
Check if test spec meets the requirements.  If user does not specify any
valid test paths neither from the test spec nor from the command line,
report error.  If 'devices' property is not specified, report error.  If
no device spec is specified, report error.  If screen size property is
not one of 'small', 'normal', 'large' and 'xlarge', report error.  If
app-root is not specified or is not a directory, report error.  If
apppath is not specified or is not a file, report error.

Change-Id: I37a6f82aa926ab09d417f77c0eb8948c60907b72