feat(mdtest create): mdtest helps users to create a template test spec and provide relevant instructions on how to create valid spec from the template
mdtest now support "mdtest create" command which is used to create
either a spec or a test template for the user.  --spec-template option
is added to specify the path to the template spec;  --test-template
option is added to specify the path to the template test.

mdtest will print out information regarding to each attribute in the
test spec and provide instructions to make the template a real test
spec.  mdtest will also print out some guidance for creating test
scripts and the test template is detailed with more instructions.

Now mdtest return 1 if some test fails.

Change-Id: Ie88635c2274bd6acf6c25afa89b281d8cdeb4652