feat(mdtest doctor): mdtest is able to detect if any dependent tool is not installed and provide tool installation instructions
mdtest provide doctor command which checks if Dart, Pub, Flutter, ADB,
lcov are installed and reports error as well as installation instruction
if any of those tools is not installed.  The doctor command also checks
if mobiledevice tool is installed if mdtest is running on Mac OS.
mobiledevice is a commandline tool similar to ADB but does not provide
as many features as ADB does.  However, it is the best commandline tool
that meets mdtest's need.  genhtml is used to generate code coverage
report and comes along with lcov.

Change-Id: Iea4c7d3227d3c9bca9f7b10025b7cb54e94ec617