refactor(mdtest examples): refactor mdtest file structure and add two examples for mdtest
Restructure the original mdtest and split the mdtest driver api with the
mdtest_tools to make things more clear.  Add a examples folder that
contains a chat app and a shared-counter app where both apps can be
automated by mdtest.  After this refactoring, mdtest can be easily
published on pub.

Change-Id: Id256b86a47f97662c3786b291de949c1c8059d80